Hardware Interrupts
The game should start with a static screen and only when the user presses any key, the animation should start, i.e. the car should start moving.

By default the red car should move forward in the centre lane.
In this phase, you will implement the movement of the red  car.
The left cursor key shall move the car to the left lane. If the car is already in the left most lane, it should stay there.
The right cursor key shall move the car to the right lane. If the car is already in the right most lane, it should stay there.
The up cursor key shall move the car up one block a time.
The down cursor key shall move the car down one block a time. If the car's lower part already touches the lower end of the screen, the car should not go any lower.

On pressing the escape key, the game should pause and show a small confirmation screen in the middle of the screen asking "Do you want to quit?".
On pressing y during the confirmation screen, the game should quit SMOOTHLY and go back to the main screen.
On pressing n, the screen should come back without any error or fault screen and resume SMOOTHLY.
Without the confirmation screen, y and n should have no effect on the game.
Also, on pressing the esc key again, the game should resume without any error.